Automated PC Combat 1.0
Finished Player Combat Automation
A New hook that listens to the roll messages will now process the entire thing, from checking range, ammunition, thrown weapons, miss, hit, armor and deal damage as well as process Critical Hits.

Added a new feature to allow players to consume items from the Equipments tab.
Players will be prompted with a dialog and decide wether or not they want to consume an unit of that item when using it or not. Consuming it reduces the count (much like the ammunition when attacking with a weapon that requires ammo).

Fixed a few bugs that caused memory leaks and Max Call Stack exceptions.

Optimized a lot of content from the previous Active Effects Automation commit