v1.2 - Audio Assets and Bugfixes
Fixed a bug that prevented attacks from piercing armor when properly tagged.
Fixed a bug during the rendering of PC sheets that would infinitely add the consume-item button.
Removed an unused call to the "defend" hook and added a safer way to prevent the repetition of casting abilities and attacks when the GM selected a token that just rolled.